backend-server/frog_project/api_backend/requirements.txt
flask
flask_sqlalchemy
Flask-Login
flask-apscheduler
bcrypt
gunicorn
gevent
psycopg2
Werkzeug
requests